perf(recon): narrow subpel MC intermediates to 16-bit lanes
Feature ID: RECON-SUBPEL-MC.

AV2 § 7.13.3.16 fixes InterRound0 at 3 and notes the horizontal filter
output always fits within 16 bits. § 6 Table 6.3 makes that unconditional
here: BitDepth is 8 or 10, and across the § 7.13.3.18 Subpel_Filters rows
the positive taps sum to at most 184 and the negative taps to at most -56,
so the pass spans -7161..=23529 at 10-bit and the zero-phase shortcut
writes sample << 4 <= 16368.

The § 7.13.3.18 `intermediate` array therefore becomes i16, halving the
vertical pass's loads, and every tap product now enters sign-extended from
i16 on both sides so the target folds the widening into one multiply-
accumulate. The BILINEAR tap pair stays wholly in 16-bit lanes: its
weighted sum is at most 16 * 1023. `ACTIVE_TAP_SPANS` pruning is unchanged.

Decoded output is bit-identical: 30 frames of the 1080p10 reference stream
hash 48f0dc140be565069838bcf7141aba3c80cefaa14400284840b2e1475a3be945 at
both 1 and 10 threads, and the corpus decoder oracle is unchanged. Paired
alternating medians against d1eb9ad0e: 1 thread -0.64% (9/11 pairs),
2 threads -1.13% (7/11), 4 threads -0.60%, 8 threads -0.55% (10/15),
10 threads -0.85% (15/21).

+++ crates/splot-decode/src/prediction/inter/mc.rs
@@ -1290,7 +1290,7 @@ fn predict_compound_average_into<T: ReconSample, O: CompoundAverageOutput>(
     params: &[SubpelPredictParams; 2],
     cwp_weight: i16,
     pred0_scratch: Option<&mut [i32]>,
-    intermediate_scratch: Option<&mut [i32]>,
+    intermediate_scratch: Option<&mut [i16]>,
     output: &mut [O],
     output_stride: usize,
 ) -> Result<()> {
@@ -1330,7 +1330,7 @@ fn predict_compound_average_pred0_into<T: ReconSample, O: CompoundAverageOutput>
     params: &[SubpelPredictParams; 2],
     cwp_weight: i16,
     pred0: &mut [i32],
-    mut intermediate_scratch: Option<&mut [i32]>,
+    mut intermediate_scratch: Option<&mut [i16]>,
     output: &mut [O],
     output_stride: usize,
 ) -> Result<()> {
